requirements.txt
altgraph==0.17.2
Bottleneck==1.3.2
certifi==2021.10.8
charset-normalizer==2.0.7
cycler==0.11.0
future==0.18.2
idna==3.3
kiwisolver==1.3.2
matplotlib==3.4.3
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
numexpr==2.7.3
numpy==1.21.2
olefile==0.46
pandas==1.3.4
pefile==2021.9.3
Pillow==8.4.0
pip==21.2.4
plotly==5.3.1
psutil==5.8.0
pyinstaller==4.6
pyinstaller-hooks-contrib==2021.3
pyparsing==3.0.4
pypiwin32==223
pyqt5==5.9.1
python-dateutil==2.8.2
pytz==2021.3
pywin32==302
pywin32-ctypes==0.2.0
requests==2.26.0
scipy==1.7.2
seaborn==0.11.2
setuptools==58.0.4
sip==4.19.13
six==1.16.0
tenacity==8.0.1
urllib3==1.26.7
wheel==0.37.0
wincertstore==0.2
